Error with the variable types in the `drawLine` parser function
The points generated in `drawLine` are strings and they need to be transformed into numbers before going into the `drawLineFromPoints` function, otherwise they end up messing up all the indexes of the data array.

You also have to be careful with the value of `lineCount` function in the tests. it is OK to check that value it in the tests, but the problem never asks about that number. The correct solution is the result of the `intersectionCount` function (btw both values need to be updated in the `compute response`function on the test suit with the new values after the fix).